requirements.txt
glfw==2.7.0
numpy==1.26.4
PyOpenGL==3.1.7
Pillow==9.5.0